requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
Pillow>=10.0.0
fpdf2>=2.7.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
playwright>=1.40.0
flask>=3.0.0
gunicorn>=21.2.0
